requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
et-xmlfile==1.0.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
jdcal==1.4.1
jieba==0.42.1
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
numpy==1.19.0
openpyxl==3.0.4
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
Werkzeug==1.0.1
wordcloud==1.7.0
xlwt==1.3.0
